Rework of Unlock tx parsing.
This simplifies some parts of output parsing and lets us read
transferred amounts and public keys involved.
All outputs parsed need to have P2PKH scripts associated with them.

Input parsing now returns txo references (outpoints).

The Unlock tx format now allows inputs and outputs not involved in the
unlock itself to be present in the tx.

Output indices are now `uint32` instead of `uint16`.

+++ contracts/DogeParser/DogeMessageLibraryForTests.sol
@@ -86,7 +86,7 @@ contract DogeMessageLibraryForTests {
     }
 
     function parseLockTransaction(bytes calldata txBytes, bytes20 expected_output_public_key_hash) public pure
-             returns (uint, address, uint16) {
+             returns (uint, address, uint32) {
         return DogeMessageLibrary.parseLockTransaction(txBytes, expected_output_public_key_hash);
     }
